board-shooter & board-shooter_u: move ADSP out from the SMI region
Something somewhere only wants SMI memory in the SMI region, so move out ADSP from the SMI region. Shrink ADSP to the necessary amount of RAM for 720P video capture and reduce the sizes of surfaceflinger and framebuffer in order to make them fit before the userspace allocation. Allocate ADSP after the userspace allocation.

This leaves userspace with 861MB of RAM.

Signed-off-by: Sultanxda <sultanxda@gmail.com>