msm: audio: qdsp6v3: Modify the wait timer for apr registration
As part of subsystem restart, there was a  deadlock where
the kernel acquires a mutex as part of misc device open for
the pcm out driver and the kickstart thread also tries to
acquire the same mutex before restarting the modem
as the apr register was a blocking call.

Instead of waiting indefinitely for lpass or modem to be restarted
the apr register will wait for one second timeout and if the
subsystems are not up, return failure to userspace.
This will release the mutex acquired as part of misc device open
and kickstart thread can reload the modem.

Change-Id: I21a078314ce5fff24bb0cd43fe419dd5a2580ac5
CRs-fixed: 312818
Signed-off-by: Bharath Ramachandramurthy <bramacha@codeaurora.org>
Signed-off-by: Swaminathan Sathappan <Swami@codeaurora.org>